Faster pytorch batched version.
I made this very fast pytorch implementation of your SMPL model for my work with the SMIL model.
It took me all of monday, and then it turned out you had already made a pytorch version when I was done with my code. :)
I believe this should work with both the SMPL and SMIL model. It is very very fast on my computer, only limited by the memory, and it works with sparse tensors too! (Saving a lot of said memory)
I hope my code is not too messy, I don't have time to clean it up but I hope it is helpful to you.
Try it out!

Thank you so much for your work, it helped me a lot.
SMIL: https://www.iosb.fraunhofer.de/servlet/is/82920/

If you have any questions feel free to ask.
Best regards, Sebastian.